refactor: Обновить импорты и объединить типы
