+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/Panels/Settings/TextSettings/fileTypes.ts
@@ -1,0 +1,23 @@
+import {
+  TextPropSize,
+  TextPropView,
+  TextPropAlign,
+  TextPropWeight,
+  TextPropDisplay,
+  TextPropFont,
+  TextPropType,
+} from '@consta/uikit/Text'
+import { AlignTextType } from '../../../../coreTypes'
+
+export type ValueTypes =
+  | TextPropSize
+  | TextPropView
+  | TextPropAlign
+  | TextPropWeight
+  | TextPropDisplay
+  | TextPropFont
+  | TextPropType
+  | string
+  | TextPropType
+  | null
+  | AlignTextType
